fix: box ownership release race, safer card-box/econ handling, int conn ids
- BoxEngine/BoxAuthority: a client can no longer release or retire a box it
  never owned. Stop a joiner transmitting its initial Free mirror scan, and
  make a client yield its local hold when the host resolved a near
  simultaneous pickup to another player (instead of staying desynced).
- CardBoxOps: validate the whole card-box payload before minting and retain
  the box if the mint fails, rather than destroying it after a partial add.
- EconContrib: reject non-finite contributions and non-positive AddCoin.
- Widen wire connection ids from byte to int (Welcome.SelfId, roster and
  relay sender ids, PlayerModelEntry.Id) so a long session cannot wrap an id
  onto an existing player.
- Dispatch log: report non-retryable failures as non-retryable.
- CHANGELOG: correct the avatar/NSFW bullets to match current behavior.

+++ src/CardShopCoop/Net/Messages/CoreMessages.cs
@@ -19,7 +19,9 @@ namespace CardShopCoop.Net.Messages
 
     public sealed class RosterEntry
     {
-        public byte Id;
+        // int, not byte: host connection ids are unbounded and only wrap at int overflow.
+        // A byte here silently collided once a long session passed 255 total connections.
+        public int Id;
         public string Name;
         public ulong SteamId;
     }
@@ -153,7 +155,7 @@ namespace CardShopCoop.Net.Messages
     [NetworkMessage(MsgType.RelayTag)]
     public sealed class RelayTagMessage : INetMessage
     {
-        public byte SenderId;
+        public int SenderId;
         public byte Kind;
         public EItemType Extra;
         public MsgType Type
@@ -188,7 +190,7 @@ namespace CardShopCoop.Net.Messages
     [NetworkMessage(MsgType.RelayState, Delivery = Delivery.Transient)]
     public sealed class RelayStateMessage : INetMessage
     {
-        public byte SenderId;
+        public int SenderId;
         public PlayerStateMessage State = new PlayerStateMessage();
         public MsgType Type
         {
@@ -233,7 +235,8 @@ namespace CardShopCoop.Net.Messages
     public sealed class PlayerModelEntry
     {
         // 0 is the host; client ids are the host's canonical connection ids.
-        public byte Id;
+        // int, not byte, so ids past 255 cannot wrap onto an existing player.
+        public int Id;
         public bool Female;
         public int ModelIndex;
         public string CustomizationJson;
